misc/HtmlTestSuite/results/test-1061.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Removing render-blocking element should unblock rendering</title>
<link href="https://html.spec.whatwg.org/C/#blocking-attribute" rel="help"/>
<link href="https://html.spec.whatwg.org/C/#rendering-opportunity" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="support/test-render-blocking.js"/>

<!--
  The test is marked "optional" because even when the document is no longer
  render-blocked, the user agent is still free to take other factors, which are
  not limited by the spec, into consideration and therefore decide not to
  render. However, it is still more desirable if rendering starts
  immediately/soon.
-->

<script async="" blocking="render" class="test" data="parser-inserted async script" src="support/dummy-1.js?pipe=trickle(d1)&amp;async"/>
<script blocking="render" class="test" data="parser-inserted defer script" defer="" src="support/dummy-1.js?pipe=trickle(d1)&amp;defer"/>
<script blocking="render" class="test" data="parser-inserted module script" src="support/dummy-1.mjs?pipe=trickle(d1)" type="module"/>
<script async="" blocking="render" class="test" data="parser-inserted async module script" src="support/dummy-1.mjs?pipe=trickle(d1)&amp;async" type="module"/>

<!--
  No test for parser-inserted stylesheets and synchronous scripts because
  they are script-blocking or even parser-blocking, and they do not have new
  behaviors to test about.
-->

<script>
function addRenderBlockingElement(tag, title, attributes, optional_text) {
  let element = document.createElement(tag);
  element.className = 'test';
  element.setAttribute('data', title);
  element.blocking = 'render';
  Object.assign(element, attributes);
  if (optional_text)
    element.textContent = optional_text;
  document.head.appendChild(element);
}

addRenderBlockingElement(
    'link', 'script-inserted stylesheet link',
    {rel: 'stylesheet', blocking: 'render', href: 'support/target-red.css?pipe=trickle(d1)&amp;dynamic'});

addRenderBlockingElement(
    'script', 'script-inserted script',
    {src: 'support/dummy-1.js?pipe=trickle(d1)&amp;dynamic'});
addRenderBlockingElement(
    'script', 'script-inserted module script',
    {type: 'module', src: 'support/dummy-1.mjs?pipe=trickle(d1)&amp;dynamic'});

addRenderBlockingElement(
    'style', 'script-inserted inline style', {},
    '@import url("support/target-red.css?pipe=trickle(d1)&amp;imported&amp;dynamic")');
</script>

</head><body><div id="dummy">Some text</div>

<script>
const testElements = [...document.querySelectorAll('.test')];
const loadObservers = testElements.map(element =&gt; new LoadObserver(element));

promise_setup(async () =&gt; {
  for (let element of testElements)
    element.remove();

  // Test cases are run after rendering is unblocked.
  await new Promise(resolve =&gt; requestAnimationFrame(resolve));
});

for (let index = 0; index &lt; testElements.length; ++index) {
  promise_test(
    async () =&gt; assert_false(loadObservers[index].finished),
    'Render-blocking on ' + testElements[index].getAttribute('data') + ' is cancellable');

  // The loading can either continue or cancel. This test does not assert it.
  loadObservers[index].load.catch(() =&gt; {});
}
</script>

</body></html>
